+++ docs/src/examples/components/Attachment/Slots/AttachmentProgressExample.shorthand.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Attachment } from '@stardust-ui/react'
+
+const AttachmentProgressExampleShorthand = () => (
+  <Attachment header="Photo.jpg" action={{ icon: 'x' }} progress={33} />
+)
+
+export default AttachmentProgressExampleShorthand