+++ test/specs/components/Attachment/Attachment-test.tsx
@@ -1,0 +1,7 @@
+import { isConformant } from 'test/specs/commonTests'
+
+import Attachment from 'src/components/Attachment'
+
+describe('Attachment', () => {
+  isConformant(Attachment)
+})